Start refactoring `tasks/collection.py` (ref #1190)
